Allow deserialization of corrupt on-chain metadata.
Unfortunately, there are many mints on-chain with corrupt metadata due to a bug in the Metaplex token metadata program.

One such example is:

5UJvyhCsum71CK9ERrH8hcLw6aWzh4SjwWXToquWdWR3

Which fails to `MetadataDeserialize` with this error:

```
failed to deserialize data, err: expected bool is 0 or 1, got 122
```

As there's corrupt data after the `EditionNonce`.

With this patch, we attempt to deserialize the data into the normal `Metadata` struct.

If that fails, we try to deserialize it into the `MetadataPreV11` struct.

If that succeeds, we copy the values from the `MetadataPreV11` struct into the `Metadata` struct up to and including `EditionNonce`.

(Note: I didn't embed the `MetadataPreV11` into the `Metadata` struct as it makes initializating the `Metadata` struct too annoying.)

C.f.:

https://github.com/samuelvanderwaal/metaboss/issues/121

https://github.com/metaplex-foundation/metaplex-program-library/pull/407

https://github.com/metaplex-foundation/metaplex-program-library/blob/master/token-metadata/program/src/deser.rs#L12